# engine/openapi.yaml
openapi: 3.0.3
info:
  title: MyWant API
  description: |
    MyWant is a Go library implementing functional chain programming patterns with channels.
    This API allows you to create, manage, and execute functional processing chains (called "Wants")
    through a RESTful interface.

    ## Core Concepts

    - **Want**: A processing unit in the chain with metadata, spec, status, and state
    - **Config**: Complete execution configuration containing an array of Wants
    - **Chain Execution**: Asynchronous execution of interconnected processing units
    - **Recipe System**: Reusable component templates with parameter substitution and CRUD management
    - **State History**: All `stateValue` fields in history entries are structured objects, not strings

    ## Recipe Management

    The API now supports full CRUD operations for recipe templates:
    - Create, read, update, and delete recipe configurations
    - Recipes serve as reusable templates for generating want configurations
    - Support parameter substitution and validation
    - **Example field**: Each recipe can include a top-level `example` field containing
      pre-configured want deployments for one-click recipe deployment from the frontend UI

  version: 1.0.0
  contact:
    name: MyWant API Support
  license:
    name: MIT
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://api.mywant.dev
    description: Production server
paths:
...
  /api/v1/plugins/{filename}:
    get:
      summary: Serve a JIT-compiled card plugin as ESM
      description: |
        Compiles the matching `plugin.{jsx,tsx,js,ts}` file on demand using esbuild (JSX → ESM, ES2020 target).
        The result is cached by source MD5 so repeated requests are free.
        Plugins receive `window.React` and `window.__mywant.registerPlugin` as globals.
      operationId: servePlugin
      tags:
        - Plugins
      parameters:
        - name: filename
          in: path
          required: true
          schema:
            type: string
          description: Plugin filename, e.g. `rpg-observe.js`
          example: rpg-observe.js
      responses:
        '200':
          description: Compiled ESM plugin script
          content:
            application/javascript:
              schema:
                type: string
        '404':
          description: Plugin source not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Compilation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...